% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breast.R
\docType{data}
\name{breast}
\alias{breast}
\title{Breast Cancer Wisconsin (Original) Data Set.}
\format{
A data frame with 699 rows and 10 variables:
\describe{
\item{ID}{Sample code number: id number}
\item{clumpThickness}{Clump Thickness: 1 - 10}
\item{sizeUniformity}{Uniformity of Cell Size: 1 - 10}
\item{shapeUniformity}{Uniformity of Cell Shape: 1 - 10}
\item{marginalAdhesion}{Marginal Adhesion: 1 - 10}
\item{singleEpitheliclCellsize}{Single Epithelial Cell Size: 1 - 10}
\item{bareNuclei}{Bare Nuclei: 1 - 10}
\item{blandChromatin}{Bland Chromatin: 1 - 10}
\item{normlNucleoli}{Normal Nucleoli: 1 - 10}
\item{mitosis}{Mitoses: 1 - 10}
\item{class}{Class: (2 for benign, 4 for malignant)}
}
}
\usage{
breast
}
\description{
Original Wisconsin Breast Cancer Database.
}
\keyword{datasets}
