% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_discri.R
\name{dist_discri}
\alias{dist_discri}
\title{Distance Discrimination}
\usage{
dist_discri(train_data = NULL, x, mu, Sig, Sig_equal = FALSE)
}
\arguments{
\item{train_data}{a list containing the training data of each class.
If it is \code{NULL}, then \code{mu} and \code{Sig} must be given.}

\item{x}{a vector, a single sample data to be predicted.}

\item{mu}{a list containing true mean vector of each class. If \code{train_data}
has been given, it will be ignored.}

\item{Sig}{a list containing true covariance matrix of each class. If set
\code{Sig_equal} as \code{TRUE}, it just need to be the common covariance matrix.
If \code{train_data} has been given, it will be ignored.}

\item{Sig_equal}{logical, which indicates whether the covariance matrix of each
class is equal.}
}
\description{
This function do distance discrimination with Mahalanobis distance.
}
\examples{
library(dplyr)
library(purrr)
breast_rmna <- breast \%>\%
na.omit() \%>\%
mutate(class = as.factor(class))

train_data <- breast_rmna \%>\%
group_by(class) \%>\%
  group_split() \%>\%
  map(drop_label,label = "class")

test_data <- breast_rmna \%>\%
  mutate(class = (1:length(class))) \%>\%
  group_by(class) \%>\%
  group_split() \%>\%
  map(drop_label,label = "class")

predict_nolabel <- test_data \%>\%
  map_dbl(dist_discri,train_data = train_data,Sig_equal = FALSE)
predict1 <- ifelse(predict_nolabel == 1,2,4) \%>\% as.factor()
caret::confusionMatrix(predict1,breast_rmna$class)
}
