% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_label}
\alias{drop_label}
\title{Drop population-distinguishing variable}
\usage{
drop_label(mydata, label)
}
\arguments{
\item{mydata}{a tibble or matrix or dataframe, which contains a variable to show the
population.}

\item{label}{a character, the name of the variable to show the population.}
}
\description{
This function drop a variable in a dataframe according to its column name.
}
