% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independence_test.R
\name{independence_test}
\alias{independence_test}
\title{Independence Test}
\usage{
independence_test(X, part)
}
\arguments{
\item{X}{a matrix or dataframe, the observation data.}

\item{part}{a vector, how covariates are divided in order.}
}
\description{
This function carries out hypothesis test for independence between disjoint groups of covariates.
}
\examples{
# test whether the three variables of sweat dataset are independent.
independence_test(sweat,c(1,1,1))
}
