% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/milk.R
\docType{data}
\name{milk}
\alias{milk}
\title{Cost data for transporting milk.}
\format{
A data frame with 59 rows and 4 variables:
\describe{
\item{V1}{fuel cost}
\item{V2}{repair cost}
\item{V3}{investment capital}
\item{V4}{transporting method, "gasoline" for gasoline trucking, "diesel" for
diesel trucking}
}
}
\usage{
milk
}
\description{
A dataset containing three costs of transporting milk.
}
\keyword{datasets}
