% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnorm_mu_test.R
\name{mvnorm_multi_mu_test}
\alias{mvnorm_multi_mu_test}
\title{Multiple Population Mean Test}
\usage{
mvnorm_multi_mu_test(mydata, label)
}
\arguments{
\item{mydata}{a matrix/data frame/tibble which contains all the data with a variable to
distinguish populations.}

\item{label}{a character denotes the name of the distinguish variable.}
}
\description{
This function carries out hypothesis test for the means of two or more multivariate normal distributions.
}
\examples{
mvnorm_multi_mu_test(health,"ind")
}
