% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnorm_var_test.R
\name{mvnorm_multi_simult_test}
\alias{mvnorm_multi_simult_test}
\title{Multiple Simultaneous Test}
\usage{
mvnorm_multi_simult_test(mydata, label)
}
\arguments{
\item{mydata}{a matrix/data frame/tibble which contains all the data with a variable to
distinguish populations.}

\item{label}{a character denotes the name of the distinguish variable.}
}
\description{
This function carries out simultaneous hypothesis test for the means and covariance matrices of one or two multivariate normal distributions.
}
\examples{
mvnorm_multi_simult_test(health,"ind")
}
