% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnorm_var_test.R
\name{mvnorm_multi_var_test}
\alias{mvnorm_multi_var_test}
\title{Multiple Population Covariance Matrix Test}
\usage{
mvnorm_multi_var_test(mydata, label)
}
\arguments{
\item{mydata}{a matrix/data frame/tibble which contains all the data with a variable to
distinguish populations.}

\item{label}{a character denotes the name of the distinguish variable.}
}
\description{
This function carries out hypothesis test for the covariance matrices of two or more multivariate normal distributions.
}
