% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvnorm_var_test.R
\name{mvnorm_var_test}
\alias{mvnorm_var_test}
\title{Single or Bivariate Population Covariance Matrix Test}
\usage{
mvnorm_var_test(X, Sig0_type, Sig0)
}
\arguments{
\item{X}{a matrix or a data frame which denotes a sample.}

\item{Sig0_type}{a character denotes the type of covariance matrix of null
hypothesis.}

\item{Sig0}{a matrix denotes the covariance matrix under null hypothesis.}
}
\description{
This function carries out hypothesis test for the covariance matrices of one or two multivariate normal distributions.
}
