% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pHotelling.R
\name{pHotelling}
\alias{pHotelling}
\title{Hotelling Distribution}
\usage{
pHotelling(q, p, n, lower.tail = TRUE)
}
\arguments{
\item{q}{vector of quantiles.}

\item{p}{the number of order.}

\item{n}{the degree of freedom.}

\item{lower.tail}{logical; if TRUE (default), probabilities are P(X <= x) otherwise, P(X > x).}
}
\description{
Calculate the value of the distribution function of the Hotelling distribution.
}
\examples{
pHotelling(2,2,2)
}
