% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pWilks.R
\name{pWilks}
\alias{pWilks}
\title{Wilks Distribution}
\usage{
pWilks(q, p, n1, n2, lower.tail = TRUE)
}
\arguments{
\item{q}{vector of quantiles.}

\item{p}{the number of order.}

\item{n1}{the first degree of freedom.}

\item{n2}{the second degree of freedom.}

\item{lower.tail}{logical; if TRUE (default), probabilities are P(X <= x) otherwise, P(X > x).}
}
\description{
Calculate the value or approximate value of the distribution function of the Wilks distribution.
}
\examples{
1-pWilks(0.6621,4,57,2)
}
