% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simultaneous_CI.R
\name{simultaneous_CI}
\alias{simultaneous_CI}
\title{Simultaneous Confidence Interval}
\usage{
simultaneous_CI(X, method = "B", alpha = 0.05)
}
\arguments{
\item{X}{a matrix or dataframe, the observation data.}

\item{method}{string, "B" for Bonferroni method and "S" for Scheffe method.}

\item{a}{number, the confidence level.}
}
\description{
This function estimates simultaneous confidence interval for the mean of a
multivariate normal distribution with Bonferroni method and Scheffe method.
}
