% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweat.R
\docType{data}
\name{sweat}
\alias{sweat}
\title{Sweating in 20 healthy adult women.}
\format{
A data frame with 20 rows and 3 variables:
\describe{
\item{X1}{amount of sweat}
\item{X2}{sodium content in the body}
\item{X3}{potassium content in the body}
}
}
\source{
Example 3.2.1 of Applied Multivariate Statistical Analysis edited by Gao Huixuan.
}
\usage{
sweat
}
\description{
A dataset containing the amount of sweat and other attributes of 20 healthy
adult women.
}
\keyword{datasets}
